server/requirements.txt
Django==5.1.4
djangorestframework==3.16.1
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.5.1
pandas==2.3.3
scikit-learn==1.7.2
numpy==2.3.3
requests==2.32.5
gunicorn==23.0.0
python-decouple==3.8
whitenoise==6.8.2
psycopg2-binary==2.9.13
